+++ client/pages/post/new.js
@@ -21,7 +21,7 @@ export default function PostUpload({ userInfo }) {
 
     const [cannotSubmitMessage, setCannotSubmitMessage] = useState(false);
 
-    const [currentEditingImg, setCurrentEditingImg] = useState('');
+    // const [currentEditingImg, setCurrentEditingImg] = useState('');
 
     const [postInfo, setPostInfo] = useState({
         title: '',
@@ -36,50 +36,42 @@ export default function PostUpload({ userInfo }) {
         setModalOpened(!modalOpened);
     };
 
-    const modalEditHandler = (imgFile) => {
-        // 자른 이미지 slide state에 알맞은 page에 저장
-        // 함수 대충 틀만 잡아놓음 수정 필요
-        let editedContent = slide.map((el, idx) => {
-            if (idx === index) {
-                return { ...el, img: '', imgFile: imgFile };
-            } else {
-                return el;
-            }
-        });
-        setSlide(editedContent);
-        modalHandler();
-    };
+    // const modalEditHandler = (imgFile) => {
+    //     // 자른 이미지 slide state에 알맞은 page에 저장
+    //     // 함수 대충 틀만 잡아놓음 수정 필요
+    //     let editedContent = slide.map((el, idx) => {
+    //         if (idx === index) {
+    //             return { ...el, img: '', imgFile: imgFile };
+    //         } else {
+    //             return el;
+    //         }
+    //     });
+    //     setSlide(editedContent);
+    //     modalHandler();
+    // };
 
-    const currentEditingImgHandler = (key) => (e) => {
-        // 편집할 이미지 원본은 currentEditingImg에 저장
-        setCurrentEditingImg(e.target.files[0]);
-        const image = document.getElementById('image');
-        const cropper = new Cropper(image, {
-            aspectRatio: 16 / 9,
-            crop(event) {
-                console.log(event.detail.x);
-                console.log(event.detail.y);
-                console.log(event.detail.width);
-                console.log(event.detail.height);
-                console.log(event.detail.rotate);
-                console.log(event.detail.scaleX);
-                console.log(event.detail.scaleY);
-            },
-        });
-        // const cropper = new Cropper(e.target.files[0], {
-        //     aspectRatio: 16 / 9,
-        //     crop(event) {
-        //       console.log(event.detail.x);
-        //       console.log(event.detail.y);
-        //       console.log(event.detail.width);
-        //       console.log(event.detail.height);
-        //       console.log(event.detail.rotate);
-        //       console.log(event.detail.scaleX);
-        //       console.log(event.detail.scaleY);
-        //     },
-        //   })
-        // setCurrentEditingImg(cropper);
-    };
+    // const currentEditingImgHandler = (key) => async (e) => {
+    //     // 편집할 이미지 원본은 currentEditingImg에 저장
+    //     setCurrentEditingImg(e.target.files[0]);
+    //     // document.getElementsByClassName('cropper-hide')[0]?.setAttribute('src', URL.createObjectURL(e.target.files[0]));
+    //     const image = await document.getElementById('image');
+    //     console.log(image);
+    //     const cropper = new Cropper(image, {
+    //         aspectRatio: 1 / 1,
+    //         crop(event) {
+    //             console.log(event.detail.x);
+    //             console.log(event.detail.y);
+    //             console.log(event.detail.width);
+    //             console.log(event.detail.height);
+    //             console.log(event.detail.rotate);
+    //             console.log(event.detail.scaleX);
+    //             console.log(event.detail.scaleY);
+    //         },
+    //     });
+    //     console.log(cropper);
+    //     const file = cropper.getCroppedCanvas({ maxWidth: 4096, maxHeight: 4096 });
+    //     console.log(file);
+    // };
 
     const slideTextHandler = (index, key) => (e) => {
         setCannotSubmitMessage(false);
@@ -213,12 +205,7 @@ export default function PostUpload({ userInfo }) {
                 modalHandler={modalHandler}
             />
             {modalOpened ? (
-                <ImageEditModal
-                    currentEditingImg={currentEditingImg}
-                    currentEditingImgHandler={currentEditingImgHandler}
-                    modalEditHandler={modalEditHandler}
-                    modalHandler={modalHandler}
-                />
+                <ImageEditModal currentSlide={currentSlide} slide={slide} modalHandler={modalHandler} />
             ) : (
                 ''
             )}
