+++ client/src/post/ImageEditModal.js
@@ -1,16 +1,77 @@
-export default function ImageEditModal({
-    currentEditingImg,
-    currentEditingImgHandler,
-    modalEditHandler,
-    modalHandler,
-}) {
+import React, { useState, useEffect } from 'react';
+import Cropper from 'cropperjs';
+
+import 'cropperjs/dist/cropper.css';
+
+export default function ImageEditModal({ currentSlide, slide, modalHandler }) {
+    let cropper = '';
+    useEffect(() => {
+        const image = document.getElementById('image');
+        cropper = new Cropper(image, {
+            aspectRatio: 1 / 1,
+            crop(event) {
+                console.log(event.detail.x);
+                console.log(event.detail.y);
+                console.log(event.detail.width);
+                console.log(event.detail.height);
+                console.log(event.detail.rotate);
+                console.log(event.detail.scaleX);
+                console.log(event.detail.scaleY);
+            },
+        });
+    }, []);
+    const [currentEditingImg, setCurrentEditingImg] = useState('');
+
+    const modalEditHandler = (imgFile) => {
+        // 자른 이미지 slide state에 알맞은 page에 저장
+        // 함수 대충 틀만 잡아놓음 수정 필요
+        console.log(currentSlide);
+        const file = cropper.getCroppedCanvas().toBlob((blob) => {
+            console.log(blob);
+        });
+        console.log(file);
+        // let editedContent = slide.map((el, idx) => {
+        //     if (idx === index) {
+        //         return { ...el, img: '', imgFile: imgFile };
+        //     } else {
+        //         return el;
+        //     }
+        // });
+        // setSlide(editedContent);
+        // modalHandler();
+    };
+
+    const currentEditingImgHandler = (key) => async (e) => {
+        // 편집할 이미지 원본은 currentEditingImg에 저장
+        cropper.destroy();
+        setCurrentEditingImg(e.target.files[0]);
+        // document.getElementsByClassName('cropper-hide')[0]?.setAttribute('src', URL.createObjectURL(e.target.files[0]));
+        const image = await document.getElementById('image');
+        console.log(image);
+        cropper = new Cropper(image, {
+            aspectRatio: 1 / 1,
+            crop(event) {
+                console.log(event.detail.x);
+                console.log(event.detail.y);
+                console.log(event.detail.width);
+                console.log(event.detail.height);
+                console.log(event.detail.rotate);
+                console.log(event.detail.scaleX);
+                console.log(event.detail.scaleY);
+            },
+        });
+    };
     return (
         <div className="post-upload-image-modal">
             <div className="post-upload-modal-edit-area">
                 <div className="post-upload-modal-edit-area-inner">
                     <img
                         id="image"
-                        src={currentEditingImg ? URL.createObjectURL(currentEditingImg) : '/postDefaultImage.jpg'}
+                        src={
+                            currentEditingImg.length !== 0
+                                ? URL.createObjectURL(currentEditingImg)
+                                : '/postDefaultImage.jpg'
+                        }
                         alt="target image"
                     />
                 </div>
